+++ app/_backup_root_page_before_cleanup.tsx
@@ -1,0 +1,34 @@
+/**
+ * Backup of the previously corrupted root page (before cleanup).
+ * This file is ignored by Next.js routing (not named page/layout) and all code is commented out.
+ * Use this only as a reference while rebuilding /egenkontroll.
+ */
+/*
+imported-original-start
+"use client";
+import { useEffect, useMemo, useRef, useState } from 'react';
+
+type PagedList<T> = {
+  page: number;
+  pageSize: number;
+  itemCount: number;
+  totalItemCount: number;
+  totalPages: number;
+  items: T[];
+};
+
+export default function Home() {
+  // (Original gigantic form + corrupted insertions retained below for reference)
+  // --- BEGIN ORIGINAL/ CORRUPTED CONTENT ---
+  const [projectNumber, setProjectNumber] = useState('');
+  const [installerName, setInstallerName] = useState('');
+  const [workStreet, setWorkStreet] = useState('');
+  const [workPostalCode, setWorkPostalCode] = useState('');
+  const [workCity, setWorkCity] = useState('');
+  const [installationDate, setInstallationDate] = useState('');
+  const [clientName, setClientName] = useState('');
+  // (Truncated for brevity – full content existed in corrupted file.)
+  // The working form logic will be reconstructed cleanly inside /app/egenkontroll/page.tsx
+  return null;
+}
+*/
